+++ delivery_myparcel/models/__init__.py
@@ -6,8 +6,8 @@
 #                                                                            #
 ##############################################################################
 
-from . import res_country
 from . import stock_package_type
+from . import res_country
 from . import myparcel_mixin
 from . import myparcel_insurance_price
 from .delivery_carriers import delivery_myparcel_base
@@ -15,6 +15,7 @@ from .delivery_carriers import delivery_myparcel_postnl
 from .delivery_carriers import delivery_myparcel_ups_standard
 from .delivery_carriers import delivery_myparcel_ups_express
 from .delivery_carriers import delivery_myparcel_dpd
+from .delivery_carriers import delivery_myparcel_bpost
 from .delivery_carriers import delivery_myparcel_dhl_fy
 from .delivery_carriers import delivery_myparcel_dhl_connect
 from .delivery_carriers import delivery_myparcel_dhl_europlus